+++ lib/turbostreamer.rb
@@ -8,6 +8,8 @@ class TurboStreamer
   autoload :Template, 'turbostreamer/template'
   autoload :KeyFormatter, 'turbostreamer/key_formatter'
   autoload :Errors, 'turbostreamer/errors'
+  autoload :Buffer, 'turbostreamer/buffer'
+  autoload :StreamingBuffer, 'turbostreamer/buffer'
 
   BLANK = ::Object.new
 
@@ -332,10 +334,16 @@ class TurboStreamer
   def target!
     @encoder.flush
 
-    if @encoder.output.is_a?(::StringIO)
-      @encoder.output.string
+    output = @encoder.output
+    # Hand back the buffer ActionView gave us, not our wrapper around it --
+    # Template#render only calls `to_s` on a result it recognizes as an
+    # ActionView::OutputBuffer.
+    output = output.buffer if ::TurboStreamer::Buffer === output
+
+    if output.is_a?(::StringIO)
+      output.string
     else
-      @encoder.output
+      output
     end
   end