+++ lib/turbostreamer/handler.rb
@@ -17,7 +17,7 @@ class TurboStreamer
     
     def self.call(template, source)
       # this juggling is required to keep line numbers right in the error
-      %{__already_defined = defined?(json); json||=TurboStreamer::Template.new(self, output_buffer: output_buffer || ActionView::OutputBuffer.new); #{source}
+      %{__already_defined = defined?(json); json||=TurboStreamer::Template.new(self, output_buffer: TurboStreamer::Buffer.wrap(output_buffer || ActionView::OutputBuffer.new)); #{source}
         json.target! unless (__already_defined && __already_defined != "method")}
     end