+++ lib/turbostreamer/railtie.rb
@@ -11,7 +11,6 @@ class TurboStreamer
       ActiveSupport.on_load :action_view do
         # Require turbostreamer in here so it's only loaded if needed
         require 'turbostreamer'
-        require File.expand_path('../../../ext/actionview/buffer', __FILE__)
         require File.expand_path('../../../ext/actionview/streaming_template_renderer', __FILE__)
 
         # Register Turbostreamer with Rails